requirements.txt
SQLAlchemy
matplotlib
requests
pandas
psycopg2
python-dotenv
scheduler